+++ src/helpers/formatWarpcastProfile.ts
